Fixed Offense. Fixed array nil. Added array product images.

+++ app/admin/product.rb
@@ -13,7 +13,7 @@ ActiveAdmin.register Product do
     selectable_column
     id_column
     column :reference_code
-    column (:image) { |product| image_tag(product.image.url(:thumb)) if product.image? }
+    column (:image) { |product| image_tag(product.retrieve_main_image(:thumb)) }
     column :name
     column :enabled
     column :created_at
@@ -70,16 +70,16 @@ ActiveAdmin.register Product do
       # end
 
       render partial: 'admin/products/categories', locals:
-                                                {
-                                                    products_categories: product.products_categories,
-                                                    root_category: Category.root_category
-                                                }
+                                                     {
+                                                         products_categories: product.products_categories,
+                                                         root_category: Category.root_category
+                                                     }
 
       render partial: 'admin/products/tags', locals:
-                                                     {
-                                                         products_tags: product.products_tags,
-                                                         root_tags: Tag.root_tags
-                                                     }
+                                               {
+                                                   products_tags: product.products_tags,
+                                                   root_tags: Tag.root_tags
+                                               }
     end
 
     f.inputs 'Stock' do
@@ -122,22 +122,26 @@ ActiveAdmin.register Product do
     private
 
     def update_categories
-      product = resource
-      categories = params[:product][:products_categories_ids]
-      ProductsCategory.destroy_all(['category_id NOT IN (?)', categories])
+      unless params[:product][:products_categories_ids].blank?
+        product = resource
+        categories = params[:product][:products_categories_ids]
+        ProductsCategory.destroy_all(['category_id NOT IN (?)', categories])
 
-      categories.each do |category|
-        ProductsCategory.find_or_create_by(product_id: product.id, category_id: category)
+        categories.each do |category|
+          ProductsCategory.find_or_create_by(product_id: product.id, category_id: category)
+        end
       end
     end
 
     def update_tags
-      product = resource
-      tags = params[:product][:products_tags_ids]
-      ProductsTag.destroy_all(['tag_id NOT IN (?)', tags])
+      unless params[:product][:products_tags_ids].blank?
+        product = resource
+        tags = params[:product][:products_tags_ids]
+        ProductsTag.destroy_all(['tag_id NOT IN (?)', tags])
 
-      tags.each do |tag|
-        ProductsTag.find_or_create_by(product_id: product.id, tag_id: tag)
+        tags.each do |tag|
+          ProductsTag.find_or_create_by(product_id: product.id, tag_id: tag)
+        end
       end
     end
   end
